♻️ refactor: editor 파일 구조 분리 및 주석 추가
