--- src/components/editor/index.ts
@@ -1,2 +1,0 @@
-export { Toolbar } from './Toolbar';
-export { useEditorConfig, useEditorWithContent } from './useEditorConfig';
